+++ tests/testthat/test_desplot.R
@@ -78,8 +78,10 @@ test_that("out1,out2,out1.gpar,out2.gpar", {
 })
 
 test_that("dq", {
-  desplot(oats35, block~col+row, out1=block, dq=dq)
-  desplot(oats35, block~col+row|block, out1=block, dq=dq)
+  expect_silent({
+    desplot(oats35, block~col+row, out1=block, dq=dq)
+    desplot(oats35, block~col+row|block, out1=block, dq=dq)
+  })
 })
 
 test_that("strip.cex", {
@@ -273,6 +275,8 @@ desplot(data = dat1row,
 
 # ggdesplot
 test_that("ggdesplot", {
-  ggdesplot(oats35, ~ col+row|block, cex=1, num=gen)
+  expect_silent({
+    ggdesplot(oats35, ~ col+row|block, cex=1, num=gen)
+  })
 } )